Improve gulpfile and JS parsing
Swap babel es2015 preset for the latest one, as it includes the es2016
and es2017 JS improvements/changes as well.

Note: changes to the formatting of the gulpfile were due to Prettier
reformating the whole file to match editorconfig and Prettier formatting
standard. See https://github.com/prettier/prettier for more info on what
Prettier is and what it does 🤓